Add rom filtering
This commit merges and squashes the filters branch back to main. The intent here is to filter from decompress_all on download objects, such that listing of the inner files is filtered and the inner roms or compressed files are not extracted if undesirable. The logic surrounding use of filters is a little backwards. When reading the config, a true value indicates the filter will be enabled, which further indicates that any matching roms will be REMOVED from the generated list and will NOT be extracted. This seems a little strange with the bios and good filters, but ideally these will later be handled to ensure that all of that type will be captured regardless of other matching labels.

commit cefd4a3fa23dfb416fbb22b08c50e9682919e52e
Author: Spenser Reinhardt <spenserreinhardt@gmail.com>
Date:   Sat Feb 18 19:19:01 2017 -0500

    filtering, finish filtering

commit 644123403d55fca83a71b992de255da24b925852
Author: Spenser Reinhardt <spenserreinhardt@gmail.com>
Date:   Sat Feb 18 19:16:13 2017 -0500

    config: add world filter

commit b2b6fd282e78a6e8514b70db8a7a20ab9753e0fc
Author: Spenser Reinhardt <spenserreinhardt@gmail.com>
Date:   Thu Feb 16 07:51:57 2017 -0500

    start of filters